Unit Tests 171-175
python tests for delete_pools
